+++ lib/src/ngrx-data.module.ts
@@ -26,7 +26,8 @@ import {
 } from './entity-metadata/entity-metadata';
 
 import { EntityEffects } from './effects/entity-effects';
-import { EntityServiceFactory } from './entity-service/entity.service';
+import { EntityServiceFactory } from './entity-service/entity-service-interfaces';
+import { DefaultEntityServiceFactory } from './entity-service/default-entity-service-factory';
 
 import { EntityCache } from './reducers/entity-cache';
 import { EntityCollection } from './reducers/entity-collection';
@@ -48,6 +49,7 @@ import {
 } from './reducers/constants';
 
 import { EntitySelectors } from './selectors/entity-selectors';
+import { EntitySelectorsFactory } from './selectors/entity-selectors';
 import { EntitySelectors$Factory } from './selectors/entity-selectors$';
 import {
   Pluralizer,
@@ -76,8 +78,8 @@ export interface NgrxDataModuleConfig {
     EntityDefinitionService,
     EntityDispatcherFactory,
     EntityReducerFactory,
+    EntitySelectorsFactory,
     EntitySelectors$Factory,
-    EntityServiceFactory,
     {
       provide: EntityCollectionReducerMethodsFactory,
       useClass: DefaultEntityCollectionReducerMethodsFactory
@@ -88,6 +90,7 @@ export interface NgrxDataModuleConfig {
       deps: [EntityReducerFactory],
       useFactory: createEntityReducer
     },
+    { provide: EntityServiceFactory, useClass: DefaultEntityServiceFactory },
     { provide: Logger, useClass: DefaultLogger }
   ]
 })